requirements.txt
-f /usr/share/pip-wheels
alembic==1.4.3
awesome-slugify==1.6.5
Babel==2.8.0
blinker==1.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
htmllaundry==2.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml<5.0
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector==2.1.7
passlib==1.7.2
python-dateutil
python-editor==1.0.4
pytz==2020.1
regex==2020.7.14
six
speaklater==1.3
SQLAlchemy==1.3.19
Unidecode==0.4.21
Werkzeug==1.0.1
WTForms==2.3.3
setuptools
